requirements.txt
tqdm
PILLOW
pandas
python-pptx
ttkwidgets
matplotlib
tkcalendar
image-quality
numpy
opencv-python
pytesseract